Debug console window: - Added the DebugConsoleWindow class, which represents a debug window that can be pulled up with the '~' button. - Currently, the console window is specific to the TileEngine and responds to Lua instructions that are typed into the console. - May be worth looking into putting all debug output into the console window, while also allowing for turning debug flags on and off dynamically through the console to prevent flooding (added to to-do list).
Spritesheet file changes:
- Any animation or frame that is named "untitled" will simply not be loaded. Added some notation to ensure that the name for skipped frames ('untitled' right now) remains in sync between the Ruby editor (which writes the EDS file) and the game (which loads the EDS file).
- As a result of this, unfinished EDS files can still be loaded and used; even with multiple 'untitled' frames.

Refactoring:
- Moved some functionality up from the game states into the GameState abstract class. This relieves the states themselves from having to call the GameState GUI updating and drawing functions, while also leaving a way for the GameState to easily perform general operations without having to drastically change every state.
- GameState now has advanceFrame and drawFrame methods, which each call the abstract step and draw functions declared by the GameState subclasses. Additional functionality can be added before or after these calls now, in case there is additional behaviour that's general among all states.
- Put the TileEngine input handling from the step function into its own function, just to keep things cleaner. Did the same thing for MainMenu. Then made the handleEvent function to pass the event to the GUI (or any general GameState input handling we add in the future.

GUI changes:
- Creating the DebugConsoleWindow required the addition of the AdjustingContainer from Guichan's contribs library.
- Since TextField behaviour in Guichan is not as expected, so made a TextField in EDWT to encapsulate the TextField and add the behaviour needed.
- Added scrolling to bottom for the TextBox.

NPC smooth movement:
- On a whim, added smooth NPC movement. They only move in straight lines, and there is no collision detection, but it's an improvement over the instantaneous teleporting; much more realistic.

To-do list changes:
- Added priorities to the main coding to-do list, so that it is easier to pick off high priority tasks.
- Added some more script functions we'll need in the future to the to-do list.

Warning removal:
- Removed some compile-time warnings that resulted from implicit casts from ints to floats in the Tileset and Spritesheet drawing functions.

git-svn-id: http://robothaus.org/repos/ed@73 5033be24-7f03-4d61-c009-87962669689c

+++ src/TileEngine/Tileset.cpp
@@ -72,13 +72,13 @@ void Tileset::draw(int destX, int destY, int tileNum)
    float destTop = float(destY * TileEngine::TILE_SIZE);
    float destBottom = float((destY + 1) * TileEngine::TILE_SIZE);
 
-   float tileRight = (tilesetX + 1) * TileEngine::TILE_SIZE - 1;
-   float tileBottom = (tilesetY + 1) * TileEngine::TILE_SIZE - 1;
+   float tileRight = float((tilesetX + 1) * TileEngine::TILE_SIZE - 1);
+   float tileBottom = float((tilesetY + 1) * TileEngine::TILE_SIZE - 1);
 
    float top = float(tilesetY) / height;
-   float bottom = tileBottom / (height * TileEngine::TILE_SIZE - 1);
+   float bottom = float(tileBottom) / (height * TileEngine::TILE_SIZE - 1);
    float left = float(tilesetX) / width;
-   float right = tileRight / (width * TileEngine::TILE_SIZE - 1);
+   float right = float(tileRight) / (width * TileEngine::TILE_SIZE - 1);
 
    glBindTexture(GL_TEXTURE_2D, texture);
